requirements.txt
streamlit 
pandas 
supabase-py 
requests